fix: fix product info update quantity
